add harder endboss

+++ assets/models/world.class.js
@@ -20,31 +20,31 @@ class World {
     this.draw();
     this.setWorld();
 
-    setPausableInterval(this.worldIntervallFast.bind(this), 1000/60);
+    setPausableInterval(this.worldIntervallFast.bind(this), 1000 / 60);
     setPausableInterval(this.worldIntervallSlow.bind(this), 100);
   }
 
   /** Load game framework
-   * 
+   *
    * @param {Node} canvas - document.getElementById('canvas')
    * @param {EventListener} keyboard - bind keyboard
    */
-  initGame(canvas, keyboard){
+  initGame(canvas, keyboard) {
     this.canvas = canvas;
     this.ctx = canvas.getContext("2d");
     this.keyboard = keyboard;
     this.level = initLevel(0); //generate level object
-    this.level.enemies.forEach(arrayEnemy => {
+    this.level.enemies.forEach((arrayEnemy) => {
       this.enemy.push(arrayEnemy);
     });
-    this.endboss = this.level.enemies.find(e => e instanceof Endboss);
+    this.endboss = this.level.enemies.find((e) => e instanceof Endboss);
   }
 
   /** shares variables with objects */
   setWorld() {
     this.character.world = this; //übergibt die world variablen an den Character
     this.level.world = this;
-    this.enemy.forEach(arrayEnemy => {
+    this.enemy.forEach((arrayEnemy) => {
       arrayEnemy.world = this;
     });
   }
@@ -57,29 +57,32 @@ class World {
     this.checkEndscreen();
   }
 
-  worldIntervallSlow(){
+  worldIntervallSlow() {
     this.checkThrow();
   }
 
   /** Check if character jump off or hit by enemy */
   checkCollisionsEnemies() {
     this.level.enemies.forEach((enemy) => {
-      if (this.isJumpOff(enemy)){
-          enemy.energy -= 100;
+      if (enemy instanceof Chicken || enemy instanceof BabyChicken) {
+        if (this.isJumpOff(enemy)) {
+          enemy.energy -= 20;
           this.character.jump(5);
           this.deleteObject(this.level.enemies, enemy);
         }
-      else if (this.character.isColliding(enemy)) {
+      } else if (this.character.isColliding(enemy)) {
         this.character.hit();
         this.statusbarLive.setPercentageLive(this.character.energy);
       }
     });
   }
 
-  isJumpOff(enemy){
-    return this.character.isColliding(enemy) &&
-    this.character.isAboveGround() && 
-    this.character.isFalling();
+  isJumpOff(enemy) {
+    return (
+      this.character.isColliding(enemy) &&
+      this.character.isAboveGround() &&
+      this.character.isFalling()
+    );
   }
 
   checkCollisionsCoin() {
@@ -103,57 +106,64 @@ class World {
   }
 
   /** Check bottle collision, hitting normal enemies delete them.  */
-  checkCollisionThrowBottle(){
+  checkCollisionThrowBottle() {
     this.throwableObjects.forEach((bottle) => {
       this.level.enemies.forEach((enemy) => {
-        if (bottle.isColliding(enemy)){
+        if (bottle.isColliding(enemy)) {
           enemy.hit();
           bottle.isHitEnemy = true;
           bottle.animateThrow();
-          if (enemy.energy == 0 && enemy instanceof Chicken || enemy.energy == 0 && enemy instanceof BabyChicken) {
+          if (
+            (enemy.energy == 0 && enemy instanceof Chicken) ||
+            (enemy.energy == 0 && enemy instanceof BabyChicken)
+          ) {
             this.deleteObject(this.level.enemies, enemy);
           }
           bottle.isHitEnemy = false;
         }
-        if(bottle.y > 500){
+        if (bottle.y > 500) {
           this.deleteObject(this.throwableObjects, bottle);
         }
-      })
-    })
+      });
+    });
   }
 
   /** Check throw input and generate bottle object */
   checkThrow() {
-    if (this.isThrowing() && !this.throwing && this.throwableObjects.length < 1) {
+    if (
+      this.isThrowing() &&
+      !this.throwing &&
+      this.throwableObjects.length < 1
+    ) {
       this.character.idle_countdown = 0;
       this.character.loadImage("./assets/img/2_character_pepe/2_walk/W-21.png");
       let bottle = new ThrowableObject(
         this.positionBottleStartX(),
         this.positionBottleStartY(),
         this.character.otherDirection
-        );
+      );
       throwBottleSound.play();
       this.throwableObjects.push(bottle);
       this.statusbarBottles.countBottle(-1);
       this.throwing = true;
-    }else if(!this.keyboard.THROW){
+    } else if (!this.keyboard.THROW) {
       this.throwing = false;
     }
-}
+  }
 
-  isThrowing(){
+  isThrowing() {
     return this.keyboard.THROW && this.statusbarBottles.setBottle > 0;
   }
 
-  positionBottleStartX(){
+  positionBottleStartX() {
     return this.character.x_fix + this.character.width_fix * 0.5;
   }
 
-  positionBottleStartY(){
+  positionBottleStartY() {
     return this.character.y_fix + this.character.height_fix * 0.2;
   }
 
-  deleteObject(path, object){
+  deleteObject(path, object) {
     path.splice(path.indexOf(object), 1);
   }
 
@@ -166,7 +176,7 @@ class World {
     this.ctx.translate(-this.camera_x, 0);
 
     this.drawFixCameraObjects();
-    
+
     let self = this; // this wird nicht in requestAnimationFrame erkannt, daher wird eine hilfsvariable erzeugt.
     requestAnimationFrame(function () {
       //draw() wird immer wieder aufgerufen, um animationen anzuzeigen muss canvas immer wieder gelöscht werden!!
@@ -175,8 +185,8 @@ class World {
   }
 
   /** Drawed objects are moving into the canvas */
-  drawDynamicCameraObjects(){
-    this.level.backgroundLayers.forEach( bg => this.addArrayToWorld(bg));
+  drawDynamicCameraObjects() {
+    this.level.backgroundLayers.forEach((bg) => this.addArrayToWorld(bg));
     this.addArrayToWorld(this.level.clouds);
     this.addArrayToWorld(this.level.coins);
     this.addArrayToWorld(this.level.bottles);
@@ -187,7 +197,7 @@ class World {
   }
 
   /** Drawed objects are fixed into the canvas */
-  drawFixCameraObjects(){
+  drawFixCameraObjects() {
     // HUD
     this.addToWorld(this.statusbarLive);
     this.addToWorld(this.statusbarCoins);
@@ -228,15 +238,16 @@ class World {
   }
 
   /**Check which endscreen should load. */
-  checkEndscreen(){
-    let endboss = this.enemy.find(e => e instanceof Endboss);
-    
+  checkEndscreen() {
+    let endboss = this.enemy.find((e) => e instanceof Endboss);
+
     if (this.character.energy == 0) {
       bossMusic.pause();
       looseSound.play();
-      let path = "./assets/img/9_intro_outro_screens/game_over/oh no you lost!.png";
+      let path =
+        "./assets/img/9_intro_outro_screens/game_over/oh no you lost!.png";
       this.stopGame(path);
-    }else if (endboss.energy == 0) {
+    } else if (endboss.energy == 0) {
       bossMusic.pause();
       winSound.play();
       let path = "./assets/img/9_intro_outro_screens/game_over/game over!.png";
@@ -245,16 +256,16 @@ class World {
   }
 
   /**Render endscreen & stop interval
-   * 
+   *
    * @param {string} path - endscreen img
    */
-  stopGame(path){
-    let endscreen = document.getElementById('endscreen');
-    let imgEndscreen = document.getElementById('img-endscreen');
+  stopGame(path) {
+    let endscreen = document.getElementById("endscreen");
+    let imgEndscreen = document.getElementById("img-endscreen");
 
     imgEndscreen.src = path;
     setTimeout(() => {
-      endscreen.classList.remove('d-none');
+      endscreen.classList.remove("d-none");
       stopInterval = true;
     }, 1000);
   }
